requirements_GPU.txt
face_recognition
flask~=1.1.2
flask_ngrok
pymongo~=3.11.2
pymongo[srv]
tensorflow
numpy~=1.18.5
pillow~=8.0.1
requests~=2.24.0